KOGITO-2467: Use FailSafe maven plugin for Integration Tests
