+++ kogito-travel-agency/extended/visas/pom.xml
@@ -85,14 +85,6 @@
           </execution>
         </executions>
       </plugin>
-      <plugin>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <configuration>
-          <systemProperties>
-            <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-          </systemProperties>
-        </configuration>
-      </plugin>
     </plugins>
   </build>
   <profiles>
@@ -117,22 +109,6 @@
                   <enableHttpUrlHandler>true</enableHttpUrlHandler>
                   <!-- "-H:ReflectionConfigurationFiles" is a workaround for KOGITO-1893 -->
                   <additionalBuildArgs>-H:ReflectionConfigurationFiles=../../reflection.json,--allow-incomplete-classpath</additionalBuildArgs>
-                </configuration>
-              </execution>
-            </executions>
-          </plugin>
-          <plugin>
-            <artifactId>maven-failsafe-plugin</artifactId>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>integration-test</goal>
-                  <goal>verify</goal>
-                </goals>
-                <configuration>
-                  <systemProperties>
-                    <native.image.path>${project.build.directory}/${project.build.finalName}-runner</native.image.path>
-                  </systemProperties>
                 </configuration>
               </execution>
             </executions>
