+++ onboarding-example/onboarding-springboot/pom.xml
@@ -88,6 +88,9 @@
           <name>persistence</name>
         </property>
       </activation>
+      <properties>
+        <enable.resource.infinispan>true</enable.resource.infinispan>
+      </properties>
       <dependencies>
         <dependency>
           <groupId>org.kie.kogito</groupId>
@@ -98,20 +101,6 @@
           <artifactId>infinispan-spring-boot-starter-remote</artifactId>
         </dependency>
       </dependencies>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-surefire-plugin</artifactId>
-            <configuration>
-              <systemPropertyVariables>
-                <enable.resource.infinispan>true</enable.resource.infinispan>
-                <container.image.infinispan>${container.image.infinispan}</container.image.infinispan>
-              </systemPropertyVariables>
-            </configuration>
-          </plugin>
-        </plugins>
-      </build>
     </profile>
   </profiles>
 </project>
