+++ process-quarkus-example/src/test/java/org/kie/kogito/examples/PersonProcessIT.java
@@ -46,7 +46,7 @@ import io.quarkus.test.junit.QuarkusTest;
 @QuarkusTest
 @QuarkusTestResource(value = InfinispanQuarkusTestResource.class, initArgs = {@ResourceArg(name = "conditional", value = "true")})
 @QuarkusTestResource(value = KafkaQuarkusTestResource.class, initArgs = {@ResourceArg(name = "conditional", value = "true")})
-public class PersonProcessTest {
+public class PersonProcessIT {
 
     @Inject
     @Named("persons")
