+++ process-springboot-example/pom.xml
@@ -138,6 +138,9 @@
           <name>persistence</name>
         </property>
       </activation>
+      <properties>
+        <enable.resource.infinispan>true</enable.resource.infinispan>
+      </properties>
       <dependencies>
         <dependency>
           <groupId>org.kie.kogito</groupId>
@@ -148,21 +151,6 @@
           <artifactId>infinispan-spring-boot-starter-remote</artifactId>
         </dependency>
       </dependencies>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-surefire-plugin</artifactId>
-            <configuration>
-              <systemPropertyVariables>
-                <tests.category>persistence</tests.category>
-                <enable.resource.infinispan>true</enable.resource.infinispan>
-                <container.image.infinispan>${container.image.infinispan}</container.image.infinispan>
-              </systemPropertyVariables>
-            </configuration>
-	      </plugin>
-        </plugins>
-      </build>
     </profile>
     <profile>
       <id>events</id>
@@ -171,6 +159,9 @@
           <name>events</name>
         </property>
       </activation>
+      <properties>
+        <enable.resource.kafka>true</enable.resource.kafka>
+      </properties>
       <dependencies>
         <dependency>
           <groupId>org.kie.kogito</groupId>
