+++ process-springboot-example/src/test/java/org/kie/kogito/examples/demo/PersonProcessIT.java
@@ -49,7 +49,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = DemoApplication.class)
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD) // reset spring context after each test method
-public class PersonProcessTest {
+public class PersonProcessIT {
 
     @Container
     private static final GenericContainer<?> INFINISPAN = new InfinispanContainer().enableConditional();
