+++ process-springboot-example/src/test/java/org/kie/kogito/examples/demo/PersonsRestIT.java
@@ -52,7 +52,7 @@ import io.restassured.http.ContentType;
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = DemoApplication.class)
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD) // reset spring context after each test method
-public class PersonsRestTest {
+public class PersonsRestIT {
 	
     @Container
     private static final GenericContainer<?> INFINISPAN = new InfinispanContainer().enableConditional();
