+++ serverless-workflow-service-calls-quarkus/src/test/java/org/kogito/serverless/examples/CountriesResourceTest.java
@@ -32,7 +32,6 @@ import javax.inject.Inject;
 
 
 @QuarkusTest
-@Tag("integration")
 public class CountriesResourceTest {
 
     @Inject
