+++ serverless-workflow-service-calls-quarkus/src/test/java/org/kogito/serverless/examples/CountryServiceWorkflowTest.java
@@ -34,7 +34,6 @@ import java.util.HashMap;
 import java.util.Map;
 
 @QuarkusTest
-@Tag("integration")
 public class CountryServiceWorkflowTest {
 
     @Named("jsonservicecall")
